refactor(ActionLayout): simplify code and use non index file
