+++ admin/src/components/ActionManager/index.js
@@ -1,0 +1,1 @@
+export { default } from './ActionManager';
